src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
authlib==0.15.6
    # via python-twitter-v2
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==41.0.1
    # via authlib
dataclasses-json==0.5.8
    # via python-twitter-v2
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.89.0
    # via -r requirements.in
google-auth==2.20.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
google-auth-httplib2==0.1.0
    # via google-api-python-client
googleapis-common-protos==1.59.1
    # via google-api-core
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.4
    # via requests
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
mypy-extensions==1.0.0
    # via typing-inspect
packaging==23.1
    # via marshmallow
protobuf==4.23.2
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pyparsing==3.0.9
    # via httplib2
python-dotenv==1.0.0
    # via -r requirements.in
python-twitter-v2==0.8.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   google-api-core
    #   python-twitter-v2
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
typing-extensions==4.6.3
    # via typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.16
    # via
    #   google-auth
    #   requests
